normalize translation
